Add Response#offset and the start of a test that validates it
The test is quite hard if your clock is in sync. I'm not sure the right
way to actually do this because we're using a pool of NTP servers and
ntpdate and Net::NTP could hit different servers.

I think this is good enough for now to verify that things actually
work and return reasonable results.

closes #6